// src/app/features/intraday-unreal-pnl-query/models/unreal-sum-detail.model.ts
export interface UnrealSumDetail {
  tdate: string;
  ttypename: string;
  stock: string;
  stockname: string;
  qty: string;
  price: string;
  amt: string;
  bcost: string;
  scost: string;
  snetamt: string;
  bnetamt: string;
  unreal: string;
  breakeven: string;
  fee: string;
  tax: string;
  dseq: string;
  currnm: string;
  cntd_unreal_detail: any;
}