// src/app/features/intraday-unreal-pnl-query/models/unreal-sum.model.ts
export interface UnrealSum {
  no: string;
  ttypename: string;
  stock: string;
  stocknm: string;
  bqty: string;
  sqty: string;
  breakeven: string;
  bcost: string;
  scost: string;
  avgprice: string;
  snetamt: string;
  bnetamt: string;
  lastprice: string;
  AD: string;
  ADR: string;
  unreal: string;
  urratio: string;
  currnm: string;
}